KafkaLab/dashboard/requirements.txt
streamlit
kafka-python
pandas
plotly
